Backend for dimpom

+++ controllers/movies.js
@@ -1,9 +1,10 @@
 const Movie = require('../models/movie');
 const NotFoundError = require('../errors/not-found-err');
 const CastError = require('../errors/cast-err');
+const ForbiddenError = require('../errors/forbidden-err');
 
 const getMovies = (req, res, next) => {
-  Movie.find({})
+  Movie.find({ owner: req.user._id })
     .then((movies) => res.send(movies))
     .catch(next);
 };
@@ -53,10 +54,13 @@ const deleteMovie = (req, res, next) => {
       if (!movie) {
         throw new NotFoundError(`Пользователь с id: ${req.params._id} не найден`);
       }
+      if (req.user._id !== movie.owner.toString()) {
+        throw new ForbiddenError('Нельзя удалить фильм другого пользователя');
+      }
 
       return Movie.deleteOne(movie);
     })
-    .then(() => res.status(200).send({ message: 'Карточка удалена' }))
+    .then(() => res.status(200).send({ message: 'Фильм удален' }))
     .catch((err) => {
       if (err.name === 'CastError') {
         next(new CastError('Переданы некорректные данные.'));
